refactor config and server_state from edgraph to worker pkg
