+++ dgraph/cmd/alpha/run.go
@@ -424,14 +424,14 @@ var shutdownCh chan struct{}
 func run() {
 	bindall = Alpha.Conf.GetBool("bindall")
 
-	opts := edgraph.Options{
+	opts := worker.Options{
 		BadgerTables: Alpha.Conf.GetString("badger.tables"),
 		BadgerVlog:   Alpha.Conf.GetString("badger.vlog"),
 
 		PostingDir: Alpha.Conf.GetString("postings"),
 		WALDir:     Alpha.Conf.GetString("wal"),
 
-		MutationsMode:  edgraph.AllowMutations,
+		MutationsMode:  worker.AllowMutations,
 		AuthToken:      Alpha.Conf.GetString("auth_token"),
 		AllottedMemory: Alpha.Conf.GetFloat64("lru_mb"),
 	}
@@ -456,17 +456,17 @@ func run() {
 
 	switch strings.ToLower(Alpha.Conf.GetString("mutations")) {
 	case "allow":
-		opts.MutationsMode = edgraph.AllowMutations
+		opts.MutationsMode = worker.AllowMutations
 	case "disallow":
-		opts.MutationsMode = edgraph.DisallowMutations
+		opts.MutationsMode = worker.DisallowMutations
 	case "strict":
-		opts.MutationsMode = edgraph.StrictMutations
+		opts.MutationsMode = worker.StrictMutations
 	default:
 		glog.Error("--mutations argument must be one of allow, disallow, or strict")
 		os.Exit(1)
 	}
 
-	edgraph.SetConfiguration(opts)
+	worker.SetConfiguration(opts)
 
 	ips, err := getIPsFromString(Alpha.Conf.GetString("whitelist"))
 	x.Check(err)
@@ -483,7 +483,7 @@ func run() {
 		RaftId:              cast.ToUint64(Alpha.Conf.GetString("idx")),
 		WhiteListedIPRanges: ips,
 		MaxRetries:          Alpha.Conf.GetInt("max_retries"),
-		StrictMutations:     opts.MutationsMode == edgraph.StrictMutations,
+		StrictMutations:     opts.MutationsMode == worker.StrictMutations,
 		AclEnabled:          secretFile != "",
 		SnapshotAfter:       Alpha.Conf.GetInt("snapshot_after"),
 		AbortOlderThan:      abortDur,
@@ -499,11 +499,11 @@ func run() {
 
 	glog.Infof("x.Config: %+v", x.Config)
 	glog.Infof("x.WorkerConfig: %+v", x.WorkerConfig)
-	glog.Infof("edgraph.Config: %s", edgraph.Config)
+	glog.Infof("worker.Config: %s", worker.Config)
 
-	edgraph.InitServerState()
+	worker.InitServerState()
 	defer func() {
-		edgraph.State.Dispose()
+		worker.State.Dispose()
 		glog.Info("Finished disposing server state.")
 	}()
 
@@ -520,10 +520,10 @@ func run() {
 
 	// Posting will initialize index which requires schema. Hence, initialize
 	// schema before calling posting.Init().
-	schema.Init(edgraph.State.Pstore)
-	posting.Init(edgraph.State.Pstore)
+	schema.Init(worker.State.Pstore)
+	posting.Init(worker.State.Pstore)
 	defer posting.Cleanup()
-	worker.Init(edgraph.State.Pstore)
+	worker.Init(worker.State.Pstore)
 
 	// setup shutdown os signal handler
 	sdCh := make(chan os.Signal, 3)
@@ -555,7 +555,7 @@ func run() {
 	// Setup external communication.
 	aclCloser := y.NewCloser(1)
 	go func() {
-		worker.StartRaftNodes(edgraph.State.WALstore, bindall)
+		worker.StartRaftNodes(worker.State.WALstore, bindall)
 		// initialization of the admin account can only be done after raft nodes are running
 		// and health check passes
 		edgraph.ResetAcl()
