+++ worker/config.go
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package edgraph
+package worker
 
 import (
 	"fmt"
 	"path/filepath"
 	"time"
 
-	"github.com/golang/glog"
-
 	"github.com/dgraph-io/dgraph/posting"
 	"github.com/dgraph-io/dgraph/x"
+	"github.com/golang/glog"
 )
 
 const (
@@ -90,8 +89,8 @@ func SetConfiguration(newConfig Options) {
 // MinAllottedMemory is the minimum amount of memory needed for the LRU cache.
 const MinAllottedMemory = 1024.0
 
-// availableMemory is the total size of the memory we were able to identify.
-var availableMemory int64
+// AvailableMemory is the total size of the memory we were able to identify.
+var AvailableMemory int64
 
 func (opt *Options) validate() {
 	pd, err := filepath.Abs(opt.PostingDir)
@@ -100,13 +99,13 @@ func (opt *Options) validate() {
 	x.Check(err)
 	x.AssertTruef(pd != wd, "Posting and WAL directory cannot be the same ('%s').", opt.PostingDir)
 	if opt.AllottedMemory < 0 {
-		if allottedMemory := 0.25 * float64(availableMemory); allottedMemory > MinAllottedMemory {
+		if allottedMemory := 0.25 * float64(AvailableMemory); allottedMemory > MinAllottedMemory {
 			opt.AllottedMemory = allottedMemory
 			glog.Infof(
 				"LRU memory (--lru_mb) set to %vMB, 25%% of the total RAM found (%vMB)\n"+
 					"For more information on --lru_mb please read "+
 					"https://docs.dgraph.io/deploy/#config\n",
-				opt.AllottedMemory, availableMemory)
+				opt.AllottedMemory, AvailableMemory)
 		}
 	}
 	x.AssertTruefNoTrace(opt.AllottedMemory >= MinAllottedMemory,
